Update all recipes to Minecraft 1.21.4
The previous calcite-from-diorite-amethyst-shard branch was not properly commited/the wrong file was uploaded. This branch is the correct one with the proper changes. Bog went to sleep and Snoodew has no help.

Updates all TNW Improvements recipes to work on Minecraft version 1.21.4, and adds calcite, resin brick stair, and nether brick stair recipes.